requirements.txt
alpaca-py
backtrader
pandas
yfinance
python-dotenv